DBAL4 changes
